Create download_utils.py
